Moved MoreButton inside Card component but did not include it in link to Zoop
